Move remaining rpc modules to solana-rpc
