+++ core/Cargo.toml
@@ -28,10 +28,6 @@ fs_extra = "1.2.0"
 flate2 = "1.0"
 indexmap = { version = "1.5", features = ["rayon"] }
 itertools = "0.9.0"
-jsonrpc-core = "17.0.0"
-jsonrpc-core-client = { version = "17.0.0", features = ["ipc", "ws"] }
-jsonrpc-derive = "17.0.0"
-jsonrpc-http-server = "17.0.0"
 libc = "0.2.81"
 log = "0.4.11"
 lru = "0.6.1"
@@ -43,18 +39,15 @@ rand_chacha = "0.2.2"
 rand_core = "0.6.2"
 raptorq = "1.4.2"
 rayon = "1.5.0"
-regex = "1.3.9"
 retain_mut = "0.1.2"
 serde = "1.0.126"
 serde_bytes = "0.11"
 serde_derive = "1.0.103"
-serde_json = "1.0.56"
 solana-account-decoder = { path = "../account-decoder", version = "=1.8.0" }
 solana-banks-server = { path = "../banks-server", version = "=1.8.0" }
 solana-clap-utils = { path = "../clap-utils", version = "=1.8.0" }
 solana-client = { path = "../client", version = "=1.8.0" }
 solana-config-program = { path = "../programs/config", version = "=1.8.0" }
-solana-faucet = { path = "../faucet", version = "=1.8.0" }
 solana-gossip = { path = "../gossip", version = "=1.8.0" }
 solana-ledger = { path = "../ledger", version = "=1.8.0" }
 solana-logger = { path = "../logger", version = "=1.8.0" }
@@ -71,27 +64,26 @@ solana-sdk = { path = "../sdk", version = "=1.8.0" }
 solana-frozen-abi = { path = "../frozen-abi", version = "=1.8.0" }
 solana-frozen-abi-macro = { path = "../frozen-abi/macro", version = "=1.8.0" }
 solana-stake-program = { path = "../programs/stake", version = "=1.8.0" }
-solana-storage-bigtable = { path = "../storage-bigtable", version = "=1.8.0" }
 solana-streamer = { path = "../streamer", version = "=1.8.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.8.0" }
-solana-version = { path = "../version", version = "=1.8.0" }
 solana-vote-program = { path = "../programs/vote", version = "=1.8.0" }
-spl-token-v2-0 = { package = "spl-token", version = "=3.1.0", features = ["no-entrypoint"] }
 tempfile = "3.2.0"
 thiserror = "1.0"
-tokio = { version = "1", features = ["full"] }
-tokio_02 = { version = "0.2", package = "tokio", features = ["full"] }
-tokio-util = { version = "0.3", features = ["codec"] } # This crate needs to stay in sync with tokio_02, until that dependency can be removed
 solana-rayon-threadlimit = { path = "../rayon-threadlimit", version = "=1.8.0" }
 trees = "0.2.1"
 
 [dev-dependencies]
+jsonrpc-core = "17.0.0"
+jsonrpc-core-client = { version = "17.0.0", features = ["ipc", "ws"] }
 matches = "0.1.6"
 num_cpus = "1.13.0"
 reqwest = { version = "0.11.2", default-features = false, features = ["blocking", "rustls-tls", "json"] }
+serde_json = "1.0.56"
 serial_test = "0.5.1"
+solana-version = { path = "../version", version = "=1.8.0" }
 symlink = "0.1.0"
 systemstat = "0.1.7"
+tokio_02 = { version = "0.2", package = "tokio", features = ["full"] }
 
 [build-dependencies]
 rustc_version = "0.2"
