+++ rpc/Cargo.toml
@@ -10,37 +10,49 @@ documentation = "https://docs.rs/solana-rpc"
 edition = "2018"
 
 [dependencies]
+base64 = "0.12.3"
+bincode = "1.3.1"
 bs58 = "0.4.0"
 crossbeam-channel = "0.5"
 itertools = "0.9.0"
 jsonrpc-core = "17.0.0"
 jsonrpc-core-client = { version = "17.0.0", features = ["ipc", "ws"] }
 jsonrpc-derive = "17.0.0"
+jsonrpc-http-server = "17.0.0"
 jsonrpc-pubsub = "17.0.0"
 jsonrpc-ws-server = "17.0.0"
+libc = "0.2.81"
 log = "0.4.11"
+regex = "1.3.9"
 serde = "1.0.122"
 serde_derive = "1.0.103"
 serde_json = "1.0.56"
 solana-account-decoder = { path = "../account-decoder", version = "=1.8.0" }
 solana-client = { path = "../client", version = "=1.8.0" }
+solana-faucet = { path = "../faucet", version = "=1.8.0" }
 solana-gossip = { path = "../gossip", version = "=1.8.0" }
 solana-ledger = { path = "../ledger", version = "=1.8.0" }
 solana-measure = { path = "../measure", version = "=1.8.0" }
 solana-metrics = { path = "../metrics", version = "=1.8.0" }
+solana-perf = { path = "../perf", version = "=1.8.0" }
 solana-poh = { path = "../poh", version = "=1.8.0" }
 solana-runtime = { path = "../runtime", version = "=1.8.0" }
 solana-sdk = { path = "../sdk", version = "=1.8.0" }
+solana-stake-program = { path = "../programs/stake", version = "=1.8.0" }
 solana-storage-bigtable = { path = "../storage-bigtable", version = "=1.8.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.8.0" }
+solana-version = { path = "../version", version = "=1.8.0" }
 solana-vote-program = { path = "../programs/vote", version = "=1.8.0" }
 spl-token-v2-0 = { package = "spl-token", version = "=3.1.0", features = ["no-entrypoint"] }
 tokio = { version = "1", features = ["full"] }
+tokio_02 = { version = "0.2", package = "tokio", features = ["full"] }
+tokio-util = { version = "0.3", features = ["codec"] } # This crate needs to stay in sync with tokio_02, until that dependency can be removed
 
 [dev-dependencies]
 serial_test = "0.4.0"
 solana-logger = { path = "../logger", version = "=1.8.0" }
-solana-stake-program = { path = "../programs/stake", version = "=1.8.0" }
+solana-net-utils = { path = "../net-utils", version = "=1.8.0" }
+symlink = "0.1.0"
 
 [lib]
 crate-type = ["lib"]
