+++ rpc/src/lib.rs
@@ -2,9 +2,12 @@
 pub mod max_slots;
 pub mod optimistically_confirmed_bank_tracker;
 pub mod parsed_token_accounts;
+pub mod rpc;
 pub mod rpc_completed_slots_service;
+pub mod rpc_health;
 pub mod rpc_pubsub;
 pub mod rpc_pubsub_service;
+pub mod rpc_service;
 pub mod rpc_subscriptions;
 pub mod send_transaction_service;
 pub mod transaction_status_service;
