+++ stake-monitor/Cargo.toml
@@ -20,6 +20,7 @@ solana-cli-config = { path = "../cli-config", version = "=1.8.0" }
 solana-client = { path = "../client", version = "=1.8.0" }
 solana-logger = { path = "../logger", version = "=1.8.0" }
 solana-metrics = { path = "../metrics", version = "=1.8.0" }
+solana-rpc = { path = "../rpc", version = "=1.8.0" }
 solana-sdk = { path = "../sdk", version = "=1.8.0" }
 solana-stake-program = { path = "../programs/stake", version = "=1.8.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.8.0" }
