+++ validator/src/bin/solana-test-validator.rs
@@ -9,8 +9,8 @@ use {
         },
     },
     solana_client::rpc_client::RpcClient,
-    solana_core::rpc::JsonRpcConfig,
     solana_faucet::faucet::{run_local_faucet_with_port, FAUCET_PORT},
+    solana_rpc::rpc::JsonRpcConfig,
     solana_sdk::{
         account::AccountSharedData,
         clock::Slot,
